Restricted creating a sale order to only the store attendants
